cl_max_fps now accepts float-values (e.g. 0.5 = 1 frame every two seconds). max fps = 0 now means standstill, to uncap fps, set to < 0 Added PBR effects to raytracing shader

+++ core/cengine/include/pragma/c_engine.h
@@ -90,7 +90,7 @@ public:
 	virtual void CloseConsole() override;
 	virtual bool IsConsoleOpen() const override;
 	void UseFullbrightShader(bool b);
-	uint32_t GetFPSLimit() const;
+	float GetFPSLimit() const;
 	virtual void EndGame() override;
 	virtual bool IsClientConnected() override;
 	bool IsWindowFocused() const;
